requirements/dev.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements/dev.in
#
asgiref==3.6.0
    # via django
black==23.1.0
    # via -r requirements/dev.in
build==0.10.0
    # via pip-tools
click==8.1.3
    # via
    #   black
    #   pip-tools
django==4.1.7
    # via djangorestframework
djangorestframework==3.14.0
    # via -r requirements/base.in
flake8==6.0.0
    # via -r requirements/dev.in
mccabe==0.7.0
    # via flake8
mypy-extensions==1.0.0
    # via black
packaging==23.0
    # via
    #   black
    #   build
pathspec==0.11.1
    # via black
pip-tools==6.12.3
    # via -r requirements/dev.in
platformdirs==3.1.1
    # via black
psycopg2==2.9.5
    # via -r requirements/base.in
pycodestyle==2.10.0
    # via flake8
pyflakes==3.0.1
    # via flake8
pyproject-hooks==1.0.0
    # via build
python-decouple==3.8
    # via -r requirements/base.in
pytz==2022.7.1
    # via djangorestframework
sqlparse==0.4.3
    # via django
wheel==0.40.0
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
